% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_button}
\alias{action_button}
\title{Button}
\usage{
action_button(inputId, label = NULL, variant = "default")
}
\arguments{
\item{inputId}{input id}

\item{label}{button label}

\item{variant}{one of "default", "secondary", "destructive", "outline", "ghost", "link"}
}
\description{
Button
}
