% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{updateCheckboxInput}
\alias{updateCheckboxInput}
\title{Update a checkbox input}
\usage{
updateCheckboxInput(session, inputId, value, configuration = NULL)
}
\description{
Update a checkbox input with a new value
}
