% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{updateCheckboxInput}
\alias{updateCheckboxInput}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
updateCheckboxInput(session, inputId, value, configuration = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
